+++ Assets/Prefabs/MobNPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00129d9f4af68674ca76dde45cc5fe94
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
