+++ Assets/Prefabs/MobNPC/MobNPC_Patrol.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e41938afe3a4b3a489e43f35cda0ab22
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
