+++ Assets/Prefabs/MobNPC/NPCController.prefab.meta
