+++ Assets/Scripts/NPC/PatrolNPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20ab7be0786c7cb46b1e30f454f28047
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
