+++ Assets/Scripts/NPC/TalkingNPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5251291a78d7464a8626d84ee54e55d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
